Small features and tweaks.
- Enemy spirit size now depends on enemy size.
- There is now a gap between the leader and Pikmin when using group move.
- Fixed broken shadows when two long neighboring walls are very close and nearly facing one another.
- When repairing the suit at the ship, sparkles come out. Also, added a notification telling the player to use LMB to repair.
- The mob hurting tool now has a customizable damage percentage, and the new default is 50%, instead of 20%.
- area_image_mobs now controls whether mobs, particles, etc. appear on the area image or not.
- Pikmin that are trying to reach a dismiss spot now give up after a bit.

Also, cleaned up drawing.cpp by separating the different layers into functions, and tweaked a path stop in Play.

+++ Source/source/drawing.h
@@ -50,8 +50,8 @@ void draw_fraction(
     const unsigned int needed, const ALLEGRO_COLOR &color
 );
 void draw_health(
-    const point &center, const unsigned int health,
-    const unsigned int max_health, const float radius = DEF_HEALTH_WHEEL_RADIUS,
+    const point &center, const float health,
+    const float max_health, const float radius = DEF_HEALTH_WHEEL_RADIUS,
     const bool just_chart = false
 );
 void draw_liquid(
